Support for iframes launching other iframes (#4821)
## Describe your changes

The fun continues...
We have a few use cases in our application where one iframe will launch another one. For example, the notification banner when clicked should launch a bifrost modal. Or edit profile or verify with footprint buttons.
We *could* implement these button/banner components in footprint-js but it will become very hard to iterate on this code since we will have to release new NPM versions each time even to make a small change
So the idea is to have a "primary" iframe that can trigger other iframes. Used the terms primary and secondary in code.

<!--- An in-depth description of the changes made by this PR. Technical details and possible side effects.

Example:
- before/after screenshots (if appropriate).
- Sample payloads (if it's a API change).
- Changes x feature to such that y
- Added x
- Removed y -->

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->