Write multiple DI's per workflow based on kind (#4645)
## Describe your changes

Now that we are writing a new DI per workflow, we can go a step further
and write multiple DI's per workflow to dileneate different groups of
vreqs we make at different points/for different purposes throughout the
workflow. This PR has us write a DI for the incode doc scan vreqs,
another DI for the KYC waterfall vreqs, and another DI for the one-off
WL/AM call we make in the Alpaca workflow.

## Issue: <FP-???>

## Caveats

Currently the only time DI is actually used in a read query is for the
Middesk state machine + Watchlistcheck task, both of which are
unaffected by this change.
At the moment, this change really just helps improves our data
quality/lineage and makes debugging easier. However, this will give us
the opporunity to further optimize/clean up how we retreieve vreqs. At
the moment, in a few places we query for all outstanding+completed vreqs
by querying for the latest vreq/vres by vendor_api. We could now
optimize this to instead just query for the vreq/vres associated with
the DI for that particular stage of a workflow.

## Test Plan

<!--- How did you test your changes? :) -->