[FP-5453] Support is_live for API key tenant_roles (#4863)
## Describe your changes

Now that we have separated roles for API keys from roles for dashboard members, we can support roles that exist only in sandbox mode or live mode for API keys. This is important as the scopes granted to a role are starting to reference objects (namely vault proxies) that exist only in sandbox mode or live mode.