[Fractional] Close button on modal doesn't work (#4985)
## Describe your changes

While I was doing some product QAing today and testing older integrations, I found a bug with fractional's integration on their website. The bug is: when you click X button on footprint modal it doesn't close. This is because the SDK doesn't trigger the close event when the flow is canceled. This is a bug that we fixed ~1.5 months ago on footprint-js and exists on newer versions (I believe they are on 1.0.2, and we are on 3.x versions now) . But I believe we didn't catch this on fractional's product until now.

The one way to fix this on bifrost is to conflate onClose and onCancel callbacks we get from the parent iframe and call both whenever a cancelation happens. Also noticed that we were calling cancel at the end of the flow - while it should have been close.

Complete and cancel events will trigger like:
<img width="813" alt="Screenshot 2023-08-09 at 5 02 07 PM" src="https://github.com/onefootprint/monorepo/assets/18564210/53808229-a0bf-4e63-86af-1e5afdfb1c65">
<img width="820" alt="Screenshot 2023-08-09 at 5 00 36 PM" src="https://github.com/onefootprint/monorepo/assets/18564210/5f90be0b-e320-48f5-adc9-fba3826b494f">


<!--- An in-depth description of the changes made by this PR. Technical details and possible side effects.

Example:
- before/after screenshots (if appropriate).
- Sample payloads (if it's a API change).
- Changes x feature to such that y
- Added x
- Removed y -->

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->