[FP-5534] Remove most reads from deprecated onboarding columns (#4953)
## Describe your changes

Start moving the source of truth away from Onboarding in order to deprecate them.

The general approach is to make the about-to-be-deprecated fields private so code outside the db crate can't read them. I added some shorthand methods on Onboarding (like `ob_configuration_id(wf)`) that help you to get the ob config id from the provided WF if exists, otherwise from the current onboarding. it forced us to be very deliberate in every codepath that previously read the onboarding's obc ID