[FP-5848] Start billing for PII storage in business vaults (#5356)
## Describe your changes

We had initially decided not to do this since it's an implementation detail that each user has their business data in a separate business vault.

But, as we've built out KYB more fully, businesses are much more clearly their own entity in the dashboard, and a business may even be linked to multiple users.

This also makes calculating hot vaults much simpler - otw we'd have to calculate which user vault is considered hot when a business is used :P