[FP-5838] Add basic billing for ID documents (#5359)
## Describe your changes

Just so we have _something_ on the invoices - prices will change a bit based on the customer. I have findigs' id docs as $0 since we are charging them only the $0.50 for KYC + document onboarding.

So, we'll just apply extra scrutiny to invoices that have doc scan line items

## Caveats

The timestamp we use for counting ID doc events is pretty bad - someone could start their ID doc at the end of the month but finish it at the start of the next month, so the query isn't stable