bifurcate public onboarding config and internal onboarding config types (#5276)
## Describe your changes

Previously we were using the same API wire type to represent an onboarding configuration both in the "dashboard" customer context and in the bifrost idv world. These two views into the type are actually quite different and the best security practice here is really to only reveal the minimum amount of information needed -- since publishable keys are public we want to not reveal anything about the ob config (and future playbook) that we don't have to.

This is a WIP as will need someone on frontend to coordinate with (cc @belcedogru  😄). 
FYI this currently touches 3 endpoints:

- `GET /org/onboarding_config`
- `POST /hosted/onboarding`
- `GET /hosted/onboarding/status`


I took my best guess on what was actually needed or not, so there may be a few bugs -- especially if the frontend is actually re-using the same type across onboarding and dashboard too (in which case will need similar effort to split them up 😅.)

This is not urgent but IMO important to do before we start expanding these two objects even further!