Do not attempt to scrub a vendor error that is not parseble/deserable (#5376)
## Describe your changes

Our current handling of vendor errors is a bit gnarnia and there's a few
iterations of cleanups I think we need here. This fixes one immediate
problem which is that we can only properly save a vres from a vendor
error response if that vendor error is in the form of JSON that is
deser-able into the successful struct for that vendor. This is because
we attempt to scrub the raw json response we get from them by deser'ing
it into that singular Rust struct we have for the vendor. If we get an
error that doesn't match that schema or doesn't have a json response at
all, then we will error when we try to save the vres.

The fix here is to save vres.e_response = None when either there is no
json response or the json response is not parseable/deserable.
vres.response is non-optional atm, so we hackily save an empty json for
that. But we should just make this field optional.

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->