[FP-5903] Fail workflow guard when workflow is deactivated (#5440)
## Describe your changes

Now that we are enforcing there's only one active workflow, we should also start enforcing that deactivated workflows can't make writes.

In reality, this check must be done in a locked transaction, otherwise we open ourselves up to race conditions where a write _could_ sneak in for a deactivated workflow. The risk of this is pretty low and doesn't warrant the big refactor/more difficult programming model for now.

## Caveats

A user in a bifrost session while a workflow is deactivated will see an error. The experience today is a little clunky:

https://github.com/onefootprint/monorepo/assets/21066341/a5afefb2-dbd5-4d71-878b-973e9daf3ee6

This won't really happen often, and it's tricky to solve, so I don't think worth blocking on it. There are a few other clunky cases that we might want bifrost to handle - for ex, if a token expires (after 15min), the same experience would occur. Generally, 401 error should be handled differently, but afaik our error handling in bifrost isn't always a function of the exact error returned by the backend. a fun, big project for the future :)