Fix FootprintDeviceAttestation deser issue (#5529)
## Describe your changes

I _think_ this is the easiest way to hot fix this for now given the mess
we are in with this area of the code- will definitely prioritize
cleaning up our VendorResponse abstractions :)

- Moves the device attestion vreq/vres/risk signal saving into api/core,
so we can use test_state to test it without unwinding a whole mess of
test(cfg)'s. (and it probably doesnt make a ton of sense to be in routes
anyway)
- Made a new `FootprintDeviceAttestationData` struct to hold the
`ParsedResponse` for the attestion. This makes things more in line with
what we have for our other vendors. This is sort of a synthetic vendor
response we are making with ourself as the vendor. It just has
`associated_vault_count` and a json blob of the AppleDeviceAttestation
- Repro'd our deser error in a test and the with the new
`FootprintDeviceAttestationData` approach things are passing

I dont think its really worth fixing the (2?) existing vres's we already
saved but we could

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->