Fix redo-doc (#5577)
## Describe your changes

Right now re-do doc flow is failing. When the user opens the link, we
get `Onboarding does not exist`. I believe this is because we create the
doc Workflow with `ob_configuration_id = None` and then when we try to
get obc/tenant from the auth via workflow we error. This PR amends the
/trigger logic to give the doc workflow ob_configuration_id from the
latest Kyc workflow.

cc @belcedogru 

## Issue: <FP-???>

## Caveats

`ob_configuration_id` is shown as optional on Workflow currently but it
seems like there are many footguns around our code that are effectively
expecting it to be non-null. Wonder if we can go ahead and make this
non-null or do a better job of bifurcating the cases where it makes
sense to be null.

## Test Plan

<!--- How did you test your changes? :) -->