Improved the manual capture delay time (#5605)
## Describe your changes

##### Problem
Our manual capture takes too long after the capture button is clicked.

##### Cause
Since JS is single-threaded, it puts the `onClick` event in the event
queue, and our auto-capture algorithm is running all the time. For cases
where the algorithm fails to detect a card, it tries all the graphics
params in each interval which may take quite some time (in order of
seconds if the device is slow). This keeps the `onClick` event from
being fired until the interval is completed.

##### Solution
Instead of going through all of the params in each interval, go through
them in batches of smaller batch size.

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects.

Example:
- before/after screenshots (if appropriate).
- Sample payloads (if it's a API change).
- Changes x feature to such that y
- Added x
- Removed y -->

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan
Tested the auto-capture functionality and manual capture.
<!--- How did you test your changes? :) -->