Use merged data in kyc idv flow transitions (#5664)
## Describe your changes

KYC transitions weren't checking the correct data field in conditions. 

Details: 

The state machine context doesn't get updated until after the
transition, so when `cond` is evaluated, the action hasn't been done
yet. Some transitions are affected by `event.payload` (the newest
submitted data), so this PR fixes it. Because this is an easily
overlooked detail and because it's not much overhead, for good measure
I've made all `dataSubmitted` conditions evaluate the merged data (which
includes the newly submitted data).

Belce noticed this while we were looking for the Trayd bug. 

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects.

Example:
- before/after screenshots (if appropriate).
- Sample payloads (if it's a API change).
- Changes x feature to such that y
- Added x
- Removed y -->

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->