Impl FeatureSet and new rules (#5648)
## Describe your changes

Create a structure that allows us to define Vecs of rules that are written against a Vec of footprint reason codes in order to evaluate a `RiskSignalsForDecision` struct and return a structure that has the optionally evaluated results of each Vec of rules. (the code makes sense more than that sentence does).

The goal here is to replace all of the `{vendor}Features` structures, and replace the waterfall rules code. Having a trait that all the RSGStructs share makes this RiskSignalRuleEvaluator concise

Also adds all the flat rules we currently have defined in the separate `*Features` typed rulesets
## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

no actual behavior changes, next PR in stack