[FP-6084] Add info on docs uploaded via API in `GET /entities/<>/documents` endpoint (#5725)
## Describe your changes

Coba wants to backfill id document images for their existing users. So, we have to support uploading images outside of bifrost.

## Caveats

This is a really hacky way to support "vault-only" images that are created via API rather than via bifrost. This endpoint (and the frontend on the dashboard) were only created to really support images added via bifrost that are verified with incode.

Will chat w backend team on whether we want this hacky approach or want to add better first-class support for vault-only images


## Test Plan

Will add an integration test for API-uploaded image


https://github.com/onefootprint/monorepo/assets/21066341/9b41f5bc-4183-4037-8b1c-87687b0b8063