Fix incorrect parsing of Incode errors (#5824)
## Describe your changes

We are not properly parsing Incode errors that do not have `message`
included in the response because we have `message: String` in the Serde
struct. For eg this is parsed as a successful response and user is not
prompted to upload a new image:

```
{"timestamp":1695335887367,"status":4019,"error":"Face not found.","path":"/omni/add/face/third-party"}
```

This fixes by making `message` optional. Also makes `path` optional to
be safe.

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->