Pass PiiValue into DI-specific validation (#5819)
## Describe your changes

In a future PR, we're going to start saving whether a piece of custom data was represented as a string or as a JSON object. In order to do that, PiiValue has to bleed a little bit deeper into our internal abstractions... before this PR, we would very quickly take PiiValue::Json values and serialize them into PiiStrings. Now, we start actually passing the PiiValue into the validation function. None of the validation functions are doing validation of the JSON yet, but I'll change this soon

To make things easier, we also start generated derived entries inline in the validation functions. This also has the added benefit of letting us stop front-loading some validation for derived entries :)