Replace PiiValue with PiiJsonValue (#5822)
## Describe your changes

There was previously this strange value between PiiJsonValue and PiiValue - was a string represented as `PiiJsonValue(serde_json::Value::String(s))` or `PiiString(s)`?

Technically, the former would represent `"value"` and the latter would represent `value` - but we don't really have a need for that in our model.

Now, `PiiString` is moving towards representing the raw value stored in the vault, and `PiiJsonValue` is representing any value provided in the API that will be stored in the vault