Save VaultData format (#5823)
## Describe your changes

When data is vaulted, save whether it was a raw string or a JSON-serialized string.

This will allow us to intelligently display data in the vault. This also unblocks using JSON fields in custom data keys

A future PR will use this new column to decide whether to deserialize data in the vault as JSON