[FP-5907] Make workflow mandatory on all UserOb auth sessions (#5447)
## Describe your changes

This has technically been possible for a long time, since we migrated all user onboardings to workflow. The `UserObAuth` extractor exists specifically for endpoints that can only be hit inside of onboarding (when an onboarding/workflow exists in the auth token).

## Caveats

There used to be an old codepath on the frontend that would send the email before hitting `POST /hosted/onboarding` - i believe this would break it. i'll do some investigation

## Test Plan

Testing locally, existing integration tests