[FP-6164] Support step-up inside identify machine (#5846)
## Describe your changes

In the future, the backend will start sending out hosted bifrost links
to auth tokens are "identified" (they identify and belong to a user) but
not "authenticated" (the user still needs to authenticate themselves via
SMS or biometric). So, in order to log into the vault, the user will
have to perform a "step up" as soon as they open bifrost, inside of the
identify machine.

Changes:
- The identify machine takes in an `initialAuthToken`. When provided,
the identify machine will go to a new state, `initAuthToken`. This state
will see if the token has sufficient scopes for onboarding, in which
case it will exit the identify machine. It then uses the auth token to
hit `/identify/verify` and moves on to the SMS/biometric challenge
- The identify machine no longer requires `obConfigAuth`, when an
`initialAuthToken` is provided
- The identify machine has a new state, `authTokenInvalid` - when the
initialAuthToken can't be used to identify a user, the token is just
invalid

## Caveats

The Identify logic is definitely getting complex...

The intent is that existing flows don't change at all - this only adds
support for the new authToken-requiring-step-up flow. The backend isn't
yet generating tokens like this, but we will in the future

## Test Plan

Added some tests to the idv machine and the identify machine.

Also tested locally with an authToken that requires step up. Here's a
video showing what the interface looks like when we get an auth token in
the hosted flow that requires step up:


https://github.com/onefootprint/monorepo/assets/21066341/a58ec37d-af76-439f-8ef2-e1d889929349