Run KYB entirely in POST /process instead of /authorize (#5900)
## Describe your changes

We're about to omit making the `POST /authorize` request for first-time onboardings, including first-time KYBs. So, we need to move this logic to a place that is still executed.

We've already moved the equivalent logic (sending the `Authorize` action to the workflow) for KYC into `POST /process` as well, and there was a lingering comment suggesting we should move KYB into `POST /process` as well

## Caveats

A little clunky since we may be executing two different actions on the KYB workflow now.

Also, the name Authorize doesn't really make sense anymore.

## Test Plan

Integration tests