[FP-6184] Automatically authorize workflows that are not one-click (#5883)
## Describe your changes

We've gathered that expected behavior from tenants is that when a user enters information to sign up for their site, the tenant has access to it. Without it, there's nearly nothing that can be done with an incomplete user.

The spirit of the authorize screen is to provide oauth-style consent to what a company may access. This makes a lot of sense on one-click onboardings where you already have an account - when onboarding onto a new tenant, you didn't just enter your SSN, so it may not be expected that the tenant will get access to your SSN.

On the other hand, first-time onboardings implicitly provide consent - if you're entering your phone number or SSN to onboard onto findigs, you wouldn't be surprised that findigs has access to your SSN.
 
## Test plan

Tested locally

Removed lots of expectations that an authorize requirement is handled in bifrost integration tests. The only one where it's kept is on a multi-tenant one-click onboard