Refactor obconfig code to support coba
## Describe your changes

Right now we have special logic to handle international residential addresses. This is because we don't have playbook support for configuring document types + countries.

This PR just makes it so we can define "overrides" in rust and our logic that creates the mapping of country to doc type doesn't do any of the special international logic.

This should just be temporary

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->