[1/2][FP-6204] Denormalize is_billable onto scoped_vault (#5953)
## Describe your changes

We only bill for PII storage on vaults that either (1) were created via
API or (2) have an authorized onboarding. Vaults that don't meet this
criteria have no decryptable fields for the tenant, so it feels weird to
bill them for it...

The queries for this were getting pretty slow. This creates a new column
that denormalizes whether an SV is billable for faster computation