fix incode barcode FRC (#6053)
## Describe your changes

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects.

Example:
- before/after screenshots (if appropriate).
- Sample payloads (if it's a API change).
- Changes x feature to such that y
- Added x
- Removed y -->

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->

---

<details open="true"><summary>Generated summary (powered by <a
href="https://app.graphite.dev">Graphite</a>)</summary>

> ## TL;DR
> This pull request includes changes to the backend components of the
API core, newtypes, and vendor. The changes mainly involve updates to
the test assertions and modifications to the document barcode reading
functionality.
> 
> ## What changed
> 1. In `incode_docv.rs`, the test assertion was updated to check that
the `FootprintReasonCode::DocumentBarcodeContentDoesNotMatch` is not
contained in the `frcs` variable.
> 2. In `footprint_reason_code.rs`, the note for the
`DocumentBarcodeCouldNotBeRead` enum was updated from "Document barcode
illegible" to "Document barcode could not be read".
> 3. In `incode.rs`, the import order was changed and some unnecessary
white spaces were removed. Also, the `IncodeTest` enum was updated to
include `TwoDBarcodeContent` and `Barcode2DDetected`. The test function
`test_we_added_incode_kind_to_id_doc_kind_mapping` was also updated to
improve readability.
> 
> ## How to test
> 1. Pull the changes from this PR to your local environment.
> 2. Run the test suite to ensure all tests pass with the updated
assertions.
> 3. Test the document barcode reading functionality to ensure the
updated note is displayed when the barcode could not be read.
> 
> ## Why make this change
> The changes in this PR improve the readability and maintainability of
the code. The updated test assertions ensure that the tests accurately
reflect the expected behavior of the code. The changes to the document
barcode reading functionality provide a more accurate description of the
issue when the barcode could not be read.
</details>