Support providing email in signup challenge (#6107)
## Describe your changes

As the identify flow is getting more complex with auth, we want to make it simpler to provide an email to use to intitialize the vualt. today, the client will finish calling verify and then use another endpoint to sync the email to the vault.

## Caveats

Saves email to vault in a follow-up PR