[FP-6363] Create separate public /users/<>/risk_signals API (#6239)
## Describe your changes

We should ask tenants to use this API instead of the /entities API which
should be internal-only. There's no rush to do this, but better to do
now instead of when we need to make a breaking change to risk signals as
shown on the dashboard. And new tenants can start using this API instead
of the /entities API

## Caveats

This code is just copy pasta