Remove unauthorized fields from the decrypt response (#6250)
## Describe your changes

https://onefootprint.slack.com/archives/C053U235V7G/p1697222611603189

Allows suppressing this kind of weird behavior... If you try to decrypt
a field that you didn't request in the ob config (and it exists on the
vault, maybe added by another tenant) you'll get an error. This header
allows hiding fields added by other tenants if you don't have access to
them

## Caveats

Part of me wonders if this should just be the default. Lmk what you
think