[FP-6383] Only write SV heartbeat in non-read HTTP methods (#6296)
## Describe your changes

This is made much easier to do with the work in the PR below this to
pass as much of the HTTP request around the auth extractors as we want
:)

Also moves into the more-global UserSession auth instead of the more
specific UserWf session auth

This auth class (and this branch in the auth class) is only used for
user requests during bifrost. So, we update the heartbeat any time
there's activity inside of a bifrost session.

## Caveats

This may decrease how frequently we heartbeat. We can always (1) migrate
the check_session endpoint to hit a POST API or (2) increase the
interval of how long we consider a user "in progress" after their last
heartbeat