[FP-6343] Pass challenge data to client even when twilio errors (#6285)
## Describe your changes

If twilio errors, we still asynchronously fall back to our backup SMS
vendor to try to deliver the message. As implemented today though, we
return an HTTP 400 without the challenge data, so even if our backup
vendor delivers the message successfully, the user can't enter the PIN
successfully.

This instead serializes the error inside the HTTP 200 response. The
client should still show this error if it occurs, but it should also
save the challenge data in order to allow the user to continue

## Caveats

Can't land this until the client updates to read this error in the
payload

## Test Plan

<!--- How did you test your changes? :) -->