[FP-5593] Clear all manual reviews for a SV when a new decision is made (#6343)
## Describe your changes

I punted on thinking about how workflows relate to manual reviews...
it's complex since some workflows specifically require manual review in
some cases. And manual reviews may be raised for different reasons...

Today, we had weird logic that would only sometimes mark a manual review
as complete if it were on the most recent workflow for a user. And we
would have different ways of determining if a user required manual
review

This at least now unifies the logic to always require manual review if
the scoped vault has a manual review on _any_ workflow. We'll now also
mark _all_ manual reviews as complete when a new decision is made to
prevent these dangling MRs from sticking around

https://onefootprint.slack.com/archives/C04RHCM8FU0/p1697659097777409

## Caveats

There are a few manual reviews I should manually complete in the DB
today

## Test Plan

Locally:

Before this PR: as soon as another workflow is opened for the scoped
vault, weird things start to happen:
- the user still appears in the manual review queue but the `GET
/entities` endpoint returns that manual review isn't required - this is
from different logic in how to compute if a user has a manual review
- any decision made won't clear the manual review because it's attached
to the original WF:


https://github.com/onefootprint/monorepo/assets/21066341/4c3de459-1c70-4c8b-b308-948f6ab6f0a2



After this PR:

https://github.com/onefootprint/monorepo/assets/21066341/e15b945c-b601-4178-8566-0053097d52e7